+++ envelope/utils.py
@@ -2,7 +2,7 @@ import io
 import logging
 import smtplib
 from collections import defaultdict
-from email.utils import getaddresses
+from email.utils import getaddresses, parseaddr
 from pathlib import Path
 from socket import gaierror, timeout
 
@@ -12,20 +12,36 @@ from validate_email import validate_email  # package py3-validate-email
 logger = logging.getLogger(__name__)
 
 
-class Address:
+class Address(str):
+    """
+        You can safely access the `self.name` property to access the real name and `self.address` to access the e-mail address.
+        Example: Address("John <person@example.com>") -> self.name = "John", self.address = "person@example.com"
 
-    def __init__(self, name, address):
-        self.name: str = name
-        self.address: str = address
+        Address objects are equal if their e-mail address are equal. (Their real names might differ.)
+        Address object is equal to a string if the string contains its e-mail address or the whole representation.
+        Example: "person@example.com" == Address("John <person@example.com>") == "John <person@example.com>"  # True
 
-    def __str__(self):
-        if self.name:
-            return f"{self.name} <{self.address}>"
-        return self.address
+    """
+
+    def __new__(cls, displayed_email=None, name=None, address=None):
+        if displayed_email:
+            v = parseaddr(displayed_email)
+            name, address = v[0] or name, v[1] or address
+        if name:
+            displayed_email = f"{name} <{address}>"
+        else:
+            displayed_email = address
+        instance = super().__new__(cls, displayed_email)
+        instance._name, instance._address = name, address
+        return instance
 
     def __eq__(self, other):
-        """ Addresses are equal if their e-mail addresses are equal. (Their real names might differ.) """
-        return hash(self) == hash(other)
+        """
+        Address objects are equal if their e-mail address are equal. (Their real names might differ.)
+        Address object is equal to a string if the string contains its e-mail address or the whole representation.
+        Example: "person@example.com" == Address("John <person@example.com>") == "John <person@example.com>"  # True
+        """
+        return hash(self) == hash(other) or hash(str(self)) == hash(other)
 
     def __hash__(self):
         return hash(self.address)
@@ -33,7 +49,30 @@ class Address:
     def __repr__(self):
         return self.__str__()
 
-    def get(self, name=None, address=None) -> str:
+    @property
+    def name(self):
+        return self._name
+
+    @property
+    def address(self):
+        return self._address
+
+    def get(self, name: bool = None, address: bool = None) -> str:
+        """ Return `name` and/or `address`.
+        Example:
+            e = (Envelope()
+                .to("person1@example.com")
+                .to("person1@example.com, John <person2@example.com>")
+                .to(["person3@example.com"]))
+
+            [str(x) for x in e.to()]                # ["person1@example.com", "John <person2@example.com>", "person3@example.com"]
+            [x.get(address=False) for x in e.to()]  # ["", "John", ""]
+            [x.get(name=True) for x in e.to()]      # ["person1@example.com", "John", "person3@example.com"]
+                                                    # return an address if no name given
+            [x.get(address=True) for x in e.to()]   # ["person1@example.com", "person2@example.com", "person3@example.com"]
+                                                    # addresses only
+        """
+
         if not self:
             return ""
         if name is None and address is False:
@@ -58,17 +97,17 @@ class Address:
 
     @staticmethod
     def parse(email_or_list, single=False, allow_false=False):
-        if allow_false and email_or_list is False:
+        if allow_false and email_or_list is False:  # .from_(False), .sender(False)
             return False
-        l = [Address(real_name, address) for real_name, address in getaddresses(assure_list(email_or_list))
-             if not (real_name == address == "")]
+        addresses = [Address(name=real_name, address=address) for real_name, address in getaddresses(assure_list(email_or_list))
+                     if not (real_name == address == "")]
         if single:
-            if len(l) != 1:
+            if len(addresses) != 1:
                 raise ValueError(f"Single e-mail address expected: {email_or_list}")
-            return l[0]
-        if len(l) == 0:
+            return addresses[0]
+        if len(addresses) == 0:
             raise ValueError(f"E-mail address cannot be parsed: {email_or_list}")
-        return l
+        return addresses
 
     def is_valid(self, check_mx=False):
         if not validate_email(self.address, check_mx=False):